improvements and bug fixes

+++ database/migrations/2021_04_25_214912_remove_quantity_reserved_column_from_order_products_table.php
@@ -13,9 +13,11 @@ class RemoveQuantityReservedColumnFromOrderProductsTable extends Migration
      */
     public function up()
     {
-        Schema::table('order_products', function (Blueprint $table) {
-            $table->dropColumn('quantity_reserved');
-        });
+        if (Schema::hasColumn('order_products', 'quantity_reserved')) {
+            Schema::table('order_products', function (Blueprint $table) {
+                $table->dropColumn('quantity_reserved');
+            });
+        };
     }
 
     /**
